app/requirements.txt
Django==4.2.9
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
django-prometheus==2.3.1
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-django==0.44b0
opentelemetry-instrumentation-psycopg2==0.44b0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
